Authorization done

+++ app/Http/Controllers/Admin/UserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Validation\Validator;
 
 class UserController extends Controller
@@ -27,7 +28,12 @@ class UserController extends Controller
      */
     public function create()
     {
-        return view('admin.user.create');
+        if(Gate::allows('admin',auth()->user()))
+        {
+            return view('admin.user.create');
+        }
+        session()->flash('message','You are not authorized to perform this operation.');
+        return redirect()->route('admin.dashboard');
     }
 
     /**
@@ -38,27 +44,33 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'name'=>'required',
-            'email'=>'required|email|unique:users',
-            'password'=>'required|confirmed',
-            'image'=>'mimes:jpeg,png|max:2048',
-        ]);
+        if(Gate::allows('admin',auth()->user())) {
 
-        $data = $request->except(['_token','password','image']);
-        $data['password'] = bcrypt($request->password);
 
-            if($request->hasFile('image'))
-            {
+            $request->validate([
+                'role' => 'required',
+                'name' => 'required',
+                'email' => 'required|email|unique:users',
+                'password' => 'required|confirmed',
+                'image' => 'mimes:jpeg,png|max:2048',
+            ]);
+
+            $data = $request->except(['_token', 'password', 'image']);
+            $data['password'] = bcrypt($request->password);
+
+            if ($request->hasFile('image')) {
                 $file = $request->file('image');
                 $path = 'image/user';
-                $file->move($path,$file->getClientOriginalName());
-                $data['image'] = $path.'/'.$file->getClientOriginalName();
+                $file->move($path, $file->getClientOriginalName());
+                $data['image'] = $path . '/' . $file->getClientOriginalName();
             }
 
-        User::create($data);
-        session()->flash('message','Admin created successfully');
-        return redirect()->route('user.index');
+            User::create($data);
+            session()->flash('message', 'Admin created successfully');
+            return redirect()->route('user.index');
+        }
+        session()->flash('message','You are not authorized to perform this operation.');
+        return redirect()->route('admin.dashboard');
     }
 
     /**
@@ -80,8 +92,12 @@ class UserController extends Controller
      */
     public function edit($id)
     {
-        $data['user'] = User::findOrFail($id);
-        return view('admin.user.edit',$data);
+        if(Gate::allows('admin',auth()->user())) {
+            $data['user'] = User::findOrFail($id);
+            return view('admin.user.edit', $data);
+        }
+        session()->flash('message','You are not authorized to perform this operation.');
+        return redirect()->route('admin.dashboard');
     }
 
     /**
@@ -93,28 +109,32 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $request->validate([
-            'name'=>'required',
-            'image'=>'mimes:jpeg,png|max:2048',
-        ]);
+        if(Gate::allows('admin',auth()->user())) {
+            $request->validate([
+                'role' => 'required',
+                'name' => 'required',
+                'image' => 'mimes:jpeg,png|max:2048',
+            ]);
 
-        $data['name'] = $request->name;
+            $data['role'] = $request->role;
+            $data['name'] = $request->name;
 
-        $user = User::findOrFail($id);
-        if($request->hasFile('image'))
-        {
-            $file = $request->file('image');
-            $path = 'image/user';
-            $file->move($path,$file->getClientOriginalName());
-            $data['image'] = $path.'/'.$file->getClientOriginalName();
-            if($user->image != null && file_exists($user->image))
-            {
-                unlink($user->image);
+            $user = User::findOrFail($id);
+            if ($request->hasFile('image')) {
+                $file = $request->file('image');
+                $path = 'image/user';
+                $file->move($path, $file->getClientOriginalName());
+                $data['image'] = $path . '/' . $file->getClientOriginalName();
+                if ($user->image != null && file_exists($user->image)) {
+                    unlink($user->image);
+                }
             }
+            $user->update($data);
+            session()->flash('message', 'Admin updated successfully');
+            return redirect()->route('user.index');
         }
-        $user->update($data);
-        session()->flash('message','Admin updated successfully');
-        return redirect()->route('user.index');
+        session()->flash('message','You are not authorized to perform this operation.');
+        return redirect()->route('admin.dashboard');
     }
 
     /**
@@ -125,8 +145,13 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-        User::destroy($id);
-        session()->flash('message','Admin deleted successfully');
-        return redirect()->back();
+        if(Gate::allows('admin',auth()->user())) {
+            User::destroy($id);
+            session()->flash('message', 'Admin deleted successfully');
+            return redirect()->back();
+        }
+        session()->flash('message','You are not authorized to perform this operation.');
+        return redirect()->route('admin.dashboard');
+
     }
 }
